requirements.txt
matplotlib==3.5.1
pandas==1.2.4
Pillow==9.0.1
scikit_learn==1.1.1
torch==1.13.1
torchvision==0.14.1
torchviz==0.0.2
tqdm==4.64.0
numpy==1.22.4
notebook==6.4.12